fixed misaligned navigation items
This changes the default styling for lists so that adding spacing between list items only occurs when we have _not_ set
a `class` attribute for the list. This means that we assume that if we are setting a `class` attribute on a list, this
means that we are adding extra styling to the list, and so we should not specify our default spacing between list
elements in this case.