fixed accessibility of skip link by adding `main` tag with proper `id` attribute
This fixes the accessibility of the skip link. The idea behind the skip link is that if a user is navigating with a
keyboard and the skip link is focussed, they can hit ENTER to click on the link and thereby skip over the header with
any long navigation.

However, this also means that we need to ensure that there is a valid target on the page that the user can spring to. We
do this by wrapping all of our content in a `main` element (which also improves the accessibility because it adds a
landmark for the main content) and by adding the `id` to the `main` that the link references (in this case, the link
points to `#main`, so we want to make sure that we have a `main` element with a `main` id).