add layout for navbar
This adds some rules to the CSS layout to position the navbar on desktop.
The layout is pretty much the same in every "desktop" viewport, with the
logo positioned in the upper lefthand side, and the menu items shown on the
right of the navbar. Because of this, we only need one breakpoint to specify this layout.

We can define the `grid-template-columns` of the layout to be:
`var(--navbar-margin-width) min-content 1fr var(--navbar-margin-width)` and we define
the `--navbar-margin-width` to be:
`calc((100% - var(--content-width-xl))/ 2)`

When the viewport is smaller than `--content-width-xl`, the left and right column in
the grid layout will calculate to 0 (since `100% - var(--content-width-xl)` is negative).
When the viewport is larger than `--content-width-xl`, the `--navbar-margin-width` calculate
will activate and that causes the navbar to be centered over the content area.

This also adds a few rules for centering the navigation items vertically within the navbar.