fix focus styles for navbar
When navigating with a keyboard, the focus styles were broken because the outline color is black and the background of
the navbar is black (e.g. there is no contrast between them). By setting the border color to be white, we add an extra
white ring around the focussable items to ensure that the focus ring can be seen whenever it should be visible.